reformated ArticleDetails by moving functions into separate folders
